moving files and renaming modules for better archi
