+++ ptmd/api/queries/__init__.py
@@ -3,7 +3,6 @@
 @author: D. Batista (Terazus)
 """
 
-from .create_gdrive_file import CreateGDriveFile
-from .core import get_organisms, get_chemicals, create_gdrive_file, get_organisations, register_gdrive_file
+from .core import get_organisms, get_chemicals, get_organisations
 from .users import create_user, change_password, get_me, login
-from .validate import validate_file
+from .files import validate_file, CreateGDriveFile, register_gdrive_file, create_gdrive_file
