+++ ptmd/api/queries/core.py
@@ -10,74 +10,12 @@ It includes:
 @author: D. Batista (Terazus)
 """
 from __future__ import annotations
+from flask import jsonify, Response
 
-from re import match
-
-from flask import jsonify, Response, request
-from flask_jwt_extended import get_jwt
 from sqlalchemy.orm import Session
 
 from ptmd.utils import get_session
-from ptmd.const import ALLOWED_EXPOSURE_BATCH
-from ptmd.database import Organism, Chemical, Organisation, File, User
-from ptmd.clients.gdrive import GoogleDriveConnector
-from .create_gdrive_file import CreateGDriveFile
-
-
-def create_gdrive_file() -> tuple[Response, int]:
-    """ Function to create a file in the Google Drive using the data provided by the user. Acquire data from a
-    JSON request.
-
-    :return: tuple containing a JSON response and a status code
-    """
-    session: Session = get_session()
-    try:
-        user_id = get_jwt()['sub']
-        payload: CreateGDriveFile = CreateGDriveFile()
-        response: dict[str, str] = payload.generate_file(session=session, user=user_id)
-        session.close()
-        return jsonify({"data": {'file_url': response['alternateLink']}}), 200
-    except Exception as e:
-        session.close()
-        return jsonify({"message": str(e)}), 400
-
-
-def register_gdrive_file() -> tuple[Response, int]:
-    """ Function to register an already existing file from an external Google Drive. Requires the file to
-    be 'readable by everyone with the link'.
-
-    :return: tuple containing a JSON response and a status code
-    """
-    required_fields: list[str] = ['file_id', 'batch', 'organism']
-    session = get_session()
-    try:
-        for field in required_fields:
-            if field not in request.json or request.json[field] == "":
-                raise ValueError(f'Field {field} is required.')
-
-        file_id: str = request.json['file_id']
-        user_id: int = get_jwt()['sub']
-        batch: str = request.json['batch']
-        organisation: Organisation = session.query(User).filter(User.id == user_id).first().organisation
-        connector: GoogleDriveConnector = GoogleDriveConnector()
-        if not match(ALLOWED_EXPOSURE_BATCH, batch):
-            raise ValueError(f"Batch '{batch}' has an incorrect format.")
-
-        file: File = File(gdrive_id=file_id,
-                          batch=batch,
-                          organism_name=request.json['organism'],
-                          name=connector.get_filename(file_id),
-                          organisation_name=organisation.name,
-                          user_id=user_id,
-                          session=session)
-        session.add(file)
-        session.commit()
-        msg: str = f'file {file_id} was successfully created with internal id {file.file_id}'
-        return jsonify({"data": {"message": msg, "file": file.file_id}}), 200
-    except Exception as e:
-        return jsonify({"message": str(e)}), 400
-    finally:
-        session.close()
+from ptmd.database import Organism, Chemical, Organisation
 
 
 def get_organisms() -> tuple[Response, int]:
