+++ ptmd/lib/creator/inputs2dataframes.py
@@ -18,11 +18,11 @@ from ptmd.const import (
     TIMEPOINTS_MIN, ALLOWED_VEHICLES
 )
 from ptmd.database import get_allowed_organisms, get_organism_code, get_chemical_code_mapping
-from ptmd.model.exceptions import InputTypeError, InputValueError, InputMinError, InputRangeError
-from ptmd.model.exposure_condition import ExposureCondition
+from ptmd.lib.creator.exceptions import InputTypeError, InputValueError, InputMinError, InputRangeError
+from ptmd.lib.creator.exposure_condition import ExposureCondition
 from .interfaces import InputsToDataframes as InputsToDataframesInterface
 from .dataframes import build_general_dataframe, build_sample_dataframe
-from ptmd.clients.excel import save_to_excel
+from ptmd.lib.excel import save_to_excel
 
 
 class Inputs2Dataframes(InputsToDataframesInterface):
