+++ ptmd/lib/excel/styles.py
