+++ tests/api/queries/files/test_register_file.py
@@ -47,10 +47,10 @@ class TestRegisterFile(TestCase):
             response = test_client.post('api/file/register')
         self.assertEqual(response.json, {'message': 'File added successfully.'})
 
-    @patch('ptmd.api.queries.core.get_session', return_value=session)
+    @patch('ptmd.api.queries.files.register.get_session', return_value=session)
     @patch('ptmd.api.queries.users.get_session', return_value=session)
     @patch('ptmd.api.queries.utils.get_session', return_value=session)
-    @patch('ptmd.api.queries.core.GoogleDriveConnector', return_value=MockGoogleDrive())
+    @patch('ptmd.api.queries.files.register.GoogleDriveConnector', return_value=MockGoogleDrive())
     def test_register_file(self, mock_get_session, mock_session_1, moc_session_2, mock_gdrive):
         with app.test_client() as test_client:
             login = test_client.post('/api/login', headers=HEADERS,
